feat: screen hero element
